Move makeJpegIfNeeded func to the separate file
